% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_collectionExists.R
\name{ri_collectionExists}
\alias{ri_collectionExists}
\title{Test if collection exists}
\usage{
ri_collectionExists(collection)
}
\arguments{
\item{collection}{name of the collection}
}
\value{
TRUE if collection exists, FALSE otherwise
}
\description{
This function test if a collection exists. The collection must be
given as full path of the collection.
}
