% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_getCollection.R
\name{ri_getCollection}
\alias{ri_getCollection}
\title{Get the default IRODS collection}
\usage{
ri_getCollection()
}
\description{
This function returns the default IRODS collection set with
ri_setCollection.
}
