% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_getDatadir.R
\name{ri_getDatadir}
\alias{ri_getDatadir}
\title{Get the name of the data directory}
\usage{
ri_getDatadir()
}
\value{
character string with the name of the data directory
}
\description{
Get the name of the data directory
}
\examples{
ri_getDatadir()
}
