% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_listSubCollections.R
\name{ri_listSubCollections}
\alias{ri_listSubCollections}
\title{List subcollections wihtin a collection}
\usage{
ri_listSubCollections(collection = ri_getCollection())
}
\arguments{
\item{collection}{if NULL this is the default collection, either
the name of the collection to list the dataobjects}
}
\value{
character vector with names of subcollections
}
\description{
Provides a list of all subcollections within a collection. The
collection is either the default collection or a collection
specfied by argument
}
\details{
see \code{\link{ri_setCollection}} for how to set the default
collection.
}
