% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaAdd.R
\name{ri_metaAdd}
\alias{ri_metaAdd}
\title{Adds meta data (AVU triple) to iRODS data object}
\usage{
ri_metaAdd(
  object,
  attribute,
  value,
  collection = ri_getCollection(),
  unit = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{object}{name of data object}

\item{attribute}{name of attribute}

\item{value}{value of attribute}

\item{collection}{iRODS collection where data object resides}

\item{unit}{unit of attribute, will be NULL if not provided}

\item{overwrite}{should attribute be overwritting if allready exists

iRODS uses AVU triples (attribute - value - unit) to store meta
data to data objects, collections users etc. This function can be
used to add meta data to data objects. For this function attribute
and value are obligatory. If no unit is provided, it defaults to
NULL.

If the attribute allready exists for a data object, this function
will result in an error, unless overwrite=TRUE

If no collection argument is provided, the default collection is
assumed (see ri_setCollection).}
}
\description{
Adds meta data to an iRODS data object. Meta data consists of
attribute - value - unit triples (AVU triples). This function only
adds meta data to data objects. Adding meta data to collections,
users etc. is not possible.
}
