% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaAddColl.R
\name{ri_metaAddColl}
\alias{ri_metaAddColl}
\title{Add a metadata attribute-value-unit (AVU) triplet to a collection
in iRODS.}
\usage{
ri_metaAddColl(collection, attribute, value, units = NULL)
}
\arguments{
\item{collection}{The name of the collection to which the metadata
will be added.}

\item{attribute}{The name of the metadata attribute to be added.}

\item{value}{The value of the metadata attribute to be added.}

\item{units}{An optional unit for the metadata attribute. If not provided, it defaults to NULL.}
}
\value{
This function does not return any value. It adds the specified
metadata to the specified collection in iRODS.
}
\description{
This function adds a metadata attribute-value-unit (AVU) triplet to a collection
in iRODS. The AVU triplet consists of an attribute name, a value,
and an optional unit.
}
\details{
The function checks if the provided collection, attribute, and value
are character strings. If the units parameter is provided, it also checks
if it is a character string. If any of these checks fail, an error is
raised. Additionally, the function checks if the specified collection
exists in iRODS. If the collection does not exist, an error is raised.
}
