% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaAttExists.R
\name{ri_metaAttExists}
\alias{ri_metaAttExists}
\title{Check if meta data attribute allready exists}
\usage{
ri_metaAttExists(object, attribute, collection = ri_getCollection())
}
\arguments{
\item{object}{name of data object}

\item{attribute}{name of attribute}

\item{collection}{iRODS collection where data object resides}
}
\description{
Checks if an attribute value allready exists in the meta data (AVU
triples) of an iRODS data object
}
