% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaExists.R
\name{ri_metaExists}
\alias{ri_metaExists}
\title{Check if meta data exists}
\usage{
ri_metaExists(
  object,
  attribute,
  value,
  units = NULL,
  collection = ri_getCollection()
)
}
\arguments{
\item{object}{name of data object}

\item{attribute}{name of attribute}

\item{value}{value of attribute}

\item{units}{unit of attribute, will be NULL if not provided}

\item{collection}{iRODS collection where data object resides, if not provided, the default collection will be used}
}
\value{
TRUE if AVU tripple exists, FALSE otherwise
}
\description{
Checks if an AVU triple allready exists in the meta data
of an iRODS data object
}
