% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaExistsColl.R
\name{ri_metaExistsColl}
\alias{ri_metaExistsColl}
\title{Check if a metadata attribute exists for a collection}
\usage{
ri_metaExistsColl(collection, attribute, value, units = NULL)
}
\arguments{
\item{collection}{name of the iRODS collection}

\item{attribute}{name of the metadata attribute}

\item{value}{value of the metadata attribute}

\item{units}{(Optional) unit of the metadata attribute, will be NULL if not provided}
}
\value{
TRUE if the AVU triple exists, FALSE otherwise
}
\description{
Checks if an AVU triple already exists in the metadata of an iRODS
collection
}
\details{
This function checks if a specific Attribute-Value-Units (AVU)
triple exists in the metadata of an iRODS collection. It takes the
collection name, attribute, value, and optionally units as input
parameters. If the specified AVU triple exists in the collection's
metadata, the function returns TRUE; otherwise, it returns FALSE.
}
