% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaGetAtt.R
\name{ri_metaGetAtt}
\alias{ri_metaGetAtt}
\title{Get specific attribute from data object}
\usage{
ri_metaGetAtt(object, attribute, collection = ri_getCollection())
}
\arguments{
\item{object}{name of data object}

\item{attribute}{name of attribute}

\item{collection}{iRODS collection where data object resides

This function retrieves an specific attribute - value -unit (AVU) tripple
from an iRODS data object based on the attribute name.}
}
\description{
Gets an attribute - value -unit triple from the meta data
of an iRODS data object
}
