% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaGetColl.R
\name{ri_metaGetColl}
\alias{ri_metaGetColl}
\title{Get metadata for a collection}
\usage{
ri_metaGetColl(collection, return_list = FALSE)
}
\arguments{
\item{collection}{iRODS collection where data object resides}

\item{return_list}{Returns the metadata as list object in the
attribute "av" of the returned data.frame}
}
\value{
A data.frame with three fields for attribute, value and
units data

This function returns the meta data of an iRODS collection as
data.frame. This data.frame has 3 columns: attribute, value and
units, corresponding with the attribute, value, unit tripple of the
iRODS catalog

If return_list is TRUE an internal representation of the meta data
is returned as attribute "av" of the returned data.frame.
}
\description{
This function retrieves all meta data, i.e. AVU triples, from an
iRODS collection.
}
