% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaRemoveAtt.R
\name{ri_metaRemoveAtt}
\alias{ri_metaRemoveAtt}
\title{Remove attribute from a data object}
\usage{
ri_metaRemoveAtt(object, attribute, collection = ri_getCollection())
}
\arguments{
\item{object}{name of data object}

\item{attribute}{name of attribute}

\item{collection}{iRODS collection where data object resides

This function removes an attribute - value - unit (AVU) tripple
from an iRODS data object. Only the name of the attribute must be
supplied. It will be silently ignored if the attribute does not
exist}
}
\description{
Removes an attribute - value - unit tripple from the meta data of
an iRODS object
}
