% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaRemoveColl.R
\name{ri_metaRemoveColl}
\alias{ri_metaRemoveColl}
\title{Remove a metadata attribute-value-unit (AVU) triple from a
collection in iRODS.}
\usage{
ri_metaRemoveColl(collection, attribute, value, units = NULL)
}
\arguments{
\item{collection}{The name of the iRODS collection from which to
remove the metadata.}

\item{attribute}{The name of the metadata attribute to remove.}

\item{value}{The value of the metadata attribute to remove.}

\item{units}{(Optional) The units associated with the metadata
attribute. If not provided, it defaults to NULL.}
}
\value{
This function does not return any value. It performs the
removal operation and returns invisibly.
}
\description{
This function removes an AVU triple from the metadata of an iRODS
collection. It takes the collection name, attribute, value, and
optionally units as input parameters. If the specified AVU triple
exists in the collection's metadata, it will be removed.
}
\details{
The function first checks if the specified collection exists in
iRODS. If the collection does not exist, an error is raised. It
also validates the input parameters to ensure they are of the correct
types. If the AVU triple exists in the collection's metadata, it is
removed using the `avuRemove` function. If the AVU triple does not
exist, no action is taken.
}
