% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_put.R
\name{ri_put}
\alias{ri_put}
\title{Store data object in iRODS collection}
\usage{
ri_put(
  filename,
  collection = ri_getCollection(),
  object = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{filename}{filename or path of file to store}

\item{collection}{name of the collection where to store the data, if not provided it uses the default collection}

\item{object}{name of the data object, defaults to the basename of the path given by the file argument}

\item{overwrite}{overwrite existing data objects, default to FALSE}
}
\description{
Stores a data object in a iRODS collection. If the data object
allready exists, it is not overwritten by default
}
