% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_rm.R
\name{ri_rm}
\alias{ri_rm}
\title{Delete a data object}
\usage{
ri_rm(object, collection = ri_getCollection())
}
\arguments{
\item{object}{name of the object}

\item{collection}{collection of the object, if no collection is
provided it uses the default collection}
}
\value{
nothing

This function removes a data object from an iRODS collection. If
the data object does not exists, it will be silently ignored
}
\description{
Delete a data object from an iRODS collection
}
