% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_session.R
\name{ri_session}
\alias{ri_session}
\title{Starts an R - IRODS session}
\usage{
ri_session(env = path.expand("~/.irods/irods_environment.json"))
}
\arguments{
\item{env}{name of IRODS environment configuration}
}
\value{
nothing
}
\description{
This functions reads the IRODS environment file and starts and
IRODS session
}
