% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_setDatadir.R
\name{ri_setDatadir}
\alias{ri_setDatadir}
\title{Set the default data directory}
\usage{
ri_setDatadir(datadir)
}
\arguments{
\item{datadir}{path to the default data directory, if empty it
it uses a directory 'data' in the current working directory.}

\item{...}{additional arguments to datafile::datafileInit

Local data files are stored into the iRODS system as data objects.
The data directory is the default directory where this package
retrieves and stores the files. To manage this directory the
`datafile` package is used

 Using a seperate data directory makes it easier to seperate data
 and code. See the vignette of the `datafile` package for more
 information.}
}
\description{
This functions sets the default data directory for the user. The
directory will be created when it does not exist.
}
