% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-helpers.R
\name{ri_test_auth_file_exists}
\alias{ri_test_auth_file_exists}
\title{ricmd test function, checks if iRODS auth file exists}
\usage{
ri_test_auth_file_exists(irods_dir = path.expand("~/.irods"))
}
\arguments{
\item{irods_dir}{iRODS configuration dir, defailts ~/.irods}
}
\description{
The ri_test_... functions are meant for package testing and setting
up a demo environment. These functions are not intended for package
users.
Checks if an iRODS auth file exists
}
