% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-helpers.R
\name{ri_test_check_docker}
\alias{ri_test_check_docker}
\title{ricmd test function, check docker}
\usage{
ri_test_check_docker(verbose = FALSE)
}
\arguments{
\item{verbose}{Give verbose output

Check if Docker is installed and can be accessed without sudo
rights.

Function copied from rirods package, https://github.com/irods/irods_client_library_rirods, Martin Schobben}
}
\description{
The ri_test_... functions are meant for package testing and setting
up a demo environment. These functions are not intended for package
users.
Checks if a docker environment exists
}
