% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-helpers.R
\name{ri_test_create_auth_file}
\alias{ri_test_create_auth_file}
\title{ricmd test function, creates dummy auth file}
\usage{
ri_test_create_auth_file(irods_dir = path.expand("~/.irods"), verbose = TRUE)
}
\arguments{
\item{irods_dir}{iRODS configuration dir, defailts ~/.irods}

\item{verbose}{Be verbose

Due to some quirk in the iRODS python library, an auth file must
exist othewrwise a non interactive connection is not possible}
}
\description{
The ri_test_... functions are meant for package testing and setting
up a demo environment. These functions are not intended for package
users.
creates a dummy auth file
}
