% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-helpers.R
\name{ri_test_create_demo_env}
\alias{ri_test_create_demo_env}
\title{ricmd test function, creates demo environment file}
\usage{
ri_test_create_demo_env(irods_ip)
}
\arguments{
\item{irods_dir}{iRODS configuration dir, defailts ~/.irods}

\item{verbose}{Be verbose

This function creates an environment file to connect to the
irods_demo Docker clusters}
}
\description{
The ri_test_... functions are meant for package testing and setting
up a demo environment. These functions are not intended for package
users.
creates a demo environment file
}
