% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-helpers.R
\name{ri_test_get_catalog_provider_ip}
\alias{ri_test_get_catalog_provider_ip}
\title{ricmd test function, get catalog provider ip}
\usage{
ri_test_get_catalog_provider_ip(verbose = TRUE)
}
\arguments{
\item{verbose}{Be verbose}
}
\description{
The ri_test_... functions are meant for package testing and setting
up a demo environment. These functions are not intended for package
users.
looks up the ip of the iRODS catalog provider
}
