% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-helpers.R
\name{ri_test_valid_ip}
\alias{ri_test_valid_ip}
\title{ricmd test function, test valid ip}
\usage{
ri_test_valid_ip(ip)
}
\arguments{
\item{ip}{ip address as character}
}
\description{
The ri_test_... functions are meant for package testing and setting
up a demo environment. These functions are not intended for package
users.
Checks if a ip address is valid
}
