% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-helpers.R
\name{ri_test_valid_session}
\alias{ri_test_valid_session}
\title{ricmd test function, creates demo environment file}
\usage{
ri_test_valid_session(session)
}
\arguments{
\item{session}{iRODS session object

An valid session object contains an iRODS connection. This function
tests if a username is present for a connection. If so, the
connection is considered valid and the function returns TRUE}
}
\description{
The ri_test_... functions are meant for package testing and setting
up a demo environment. These functions are not intended for package
users.
tests if iRODS session object is valid
}
