% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_validsession.R
\name{ri_validsession}
\alias{ri_validsession}
\title{Check for a valid iRODS session}
\usage{
ri_validsession()
}
\value{
TRUE if a valid session is found, otherwise an error is
raised
}
\description{
This function checks if a valid iRODS session is present in the
ricmd environment. A session is valid if it contains 
connection info to an iRODS server.
}
