% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_collectionExists.R
\name{ri_collectionExists}
\alias{ri_collectionExists}
\title{Test if collection exists}
\usage{
ri_collectionExists(collection)
}
\arguments{
\item{collection}{name of new collection}
}
\value{
TRUE if collection exists, FALSE otherwise


This function test of a collection (subcollection) exists within an
existing collection.
}
\description{
Test if a collection allready exists
}
