% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_createCollection.R
\name{ri_createCollection}
\alias{ri_createCollection}
\title{Create a new IRODS collection}
\usage{
ri_createCollection(collection)
}
\arguments{
\item{collection}{name of new collection}
}
\value{
nothing


This function creates a new collection (subcollection) within an
existing collection.
}
\description{
Create a new collection wihtin an existing collection.
}
