% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_get.R
\name{ri_get}
\alias{ri_get}
\title{Get object from iRODS collection}
\usage{
ri_get(
  object,
  collection = ri_getCollection(),
  datadir = ri_getDatadir(),
  filename = object,
  overwrite = FALSE
)
}
\arguments{
\item{object}{name of data object}

\item{collection}{iRODS collection where data object resides}

\item{datadir}{the local directory where to store the retrieved
data object}

\item{filename}{filename of retrieved data object}

\item{overwrite}{should file be overwritting if allready exists}
}
\value{
nothing
}
\description{
Get an object from an iRODS collection. If no collection is given,
it uses the default collection.
}
\details{
This function retrieves a data object from an iRODS collection. The
default behaviour is to get the object from the default collection
(see \code{\link{ri_setCollection}})  and store it in the local data directory
(see \code{\link{ri_setDatadir}} ) using the data object's name as filename.
Using the arguments collection, datadir, and filename it is
possible to override this default behaviour.

A file is not overwritten if it allready exists in the data
directory, use the overwrite argument to force writing of the file.
If the file exists and overwrite=FALSE, then this function will
generate an error
}
