% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_listDataObjects.R
\name{ri_listDataObjects}
\alias{ri_listDataObjects}
\title{List data objects wihtin a collection}
\usage{
ri_listDataObjects(collection = ri_getCollection())
}
\arguments{
\item{collection}{name of the collection, if not provided, then
this will be the default collection.}
}
\value{
character vector with names of dataobjects
}
\description{
Provides a list of all data objects within a collection. The
collection is either the default collection or a collection
specfied by argument.
}
\details{
see \code{\link{ri_setCollection}} for how to set the default
collection.
}
