% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaGet.R
\name{ri_metaGet}
\alias{ri_metaGet}
\title{Get all meta data from an iRODS object}
\usage{
ri_metaGet(object, collection = ri_getCollection(), return_list = FALSE)
}
\arguments{
\item{object}{name of data object}

\item{collection}{iRODS collection where data object resides}

\item{return_list}{Returns the metadata as list object}
}
\value{
A data.frame with three fields for attribute, value and
units data

This function returns the meta data of an iRODS data object as
data.frame. This data.frame has 3 columns: attribute, value and
units, corresponding with the attribute, value, unit tripple of the
iRODS catalog

It return_list is TRUE an internal representation of the meta data
is returned as attribute
}
\description{
This function retrieves all meta data, i.e. AVU triples, from an
iRODS data object.
}
