% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaGet.R
\name{ri_metaGet}
\alias{ri_metaGet}
\title{Get all meta data from an iRODS object}
\usage{
ri_metaGet(object, collection = ri_getCollection())
}
\arguments{
\item{object}{name of data object}

\item{collection}{iRODS collection where data object resides}

\item{attribute}{name of attribute}
}
\value{
A data.frame with three fields for attribute, value and units data
}
\description{
This function retrieves all meta data, i.e. AVU triples, from an
iRODS data object.
}
