% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_metaRemove.R
\name{ri_metaRemove}
\alias{ri_metaRemove}
\title{Remove meta data from an object}
\usage{
ri_metaRemove(
  object,
  attribute,
  value,
  collection = ri_getCollection(),
  units = NULL
)
}
\arguments{
\item{object}{name of data object}

\item{attribute}{name of attribute}

\item{value}{value of attribute}

\item{collection}{iRODS collection where data object resides, if not provided, the default collection will be used}

\item{units}{unit of attribute, will be NULL if not provided}
}
\description{
Removes an Attribute-Value-Units (AVU) triple from an object
}
