% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_removeMeta.R
\name{ri_metaRemove}
\alias{ri_metaRemove}
\title{Remove all meta data from a data object}
\usage{
ri_metaRemove(object, collection = ri_getCollection())
}
\arguments{
\item{object}{name of data object}

\item{collection}{iRODS collection where data object resides


This function removes all AVU tripples from an iRODS object.}
}
\description{
Removes all attribute-value-unit triples from an iRODS data object
}
