% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_objectExists.R
\name{ri_objectExists}
\alias{ri_objectExists}
\title{Check if data objects exists}
\usage{
ri_objectExists(object, collection = NULL)
}
\arguments{
\item{object}{name of the object}

\item{collection}{name of the collection, if not given, it uses the
default collection}
}
\value{
TRUE of the object exists, FALSE otherwise
}
\description{
This function checks if an object exists in a collection. This
collection is either the default collection or a collection
specified by the collection argument.
}
