% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_setDatadir.R
\name{ri_setDatadir}
\alias{ri_setDatadir}
\title{Set the default data directory}
\usage{
ri_setDatadir(datadir)
}
\arguments{
\item{datadir}{path to the default data directory.}
}
\description{
This functions sets the default data directory for the user. The
directory must exists, otherwise this function generates an error
}
\details{
Local data files are stored into the iRODS system as data objects.
The data directory is the default directory where this package
retrieves and stores the files.
}
